requirements.txt
# WORKFLOW: This file defines all Python dependencies for the Trade Compliance API.
# Used by: Docker build process, local development setup, CI/CD pipelines
# Dependencies include:
# - FastAPI for web framework and API endpoints
# - SQLAlchemy for database ORM and migrations
# - Qdrant for vector search and RAG operations
# - Sentence transformers for BAAI/bge-m3 embeddings
# - Transformers for BAAI/bge-reranker-base cross-encoder
# - Scikit-learn for confidence calibration
# - Ollama for local LLM explanations
# - LangChain for RAG pipeline orchestration
# - Pandas/OpenPyXL for ETL data processing
# - Redis for caching and rate limiting
# - Testing and development tools (pytest, ruff, mypy)
# - JSON Schema validation for response integrity
#
# Installation flow: requirements.txt -> pip install -> Dependencies ready
# Build flow: Docker build -> Copy requirements.txt -> Install dependencies -> Application ready

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pgvector==0.2.4

# Vector Search & ML (Lightweight)
qdrant-client==1.15.1
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
huggingface-hub==0.19.4
# Lightweight alternatives
faiss-cpu==1.7.4
numpy==1.24.3

# LLM & RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
ollama==0.1.7

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# Caching & Async
redis==5.0.1
aioredis==2.0.1

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Security & Auth
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
ruff==0.1.6
mypy==1.7.1
black==23.11.0

# JSON Schema Validation
jsonschema==4.20.0

